+++ brd-ios/breadwallet/src/Redesign/Scenes/Swap/Main/SwapInteractor.swift
@@ -422,52 +422,4 @@ class SwapInteractor: NSObject, Interactor, SwapViewActions {
         guard let error = error else { return }
         presenter?.presentError(actionResponse: .init(error: error))
     }
-    
-//    private func fetchWkFee(for amount: Amount, completion: @escaping ((TransferFeeBasis?) -> Void)) {
-//        guard let dataStore = dataStore,
-//              let wallet = dataStore.coreSystem?.wallet(for: amount.currency),
-//              let kvStore = Backend.kvStore,
-//              let keyStore = dataStore.keyStore,
-//              let address = dataStore.address(for: amount.currency)
-//        else {
-//            presenter?.presentError(actionResponse: .init(error: SwapErrors.noFees))
-//            completion(nil)
-//            return
-//        }
-//
-//        let value = dataStore.amountFrom(decimal: amount.tokenValue, currency: amount.currency)
-//        let sender = Sender(wallet: wallet, authenticator: keyStore, kvStore: kvStore)
-//        sender.estimateFee(address: address,
-//                           amount: value,
-//                           tier: .regular,
-//                           isStake: false) { [weak self] result in
-//            switch result {
-//            case .success(let fee):
-//                completion(fee)
-//
-//            case .failure(let error):
-//                self?.presenter?.presentError(actionResponse: .init(error: error))
-//                completion(nil)
-//            }
-//        }
-//    }
-//
-//    private func fetchEthFee(for amount: Amount, completion: @escaping ((Decimal?) -> Void)) {
-//        let address = dataStore?.address(for: amount.currency)
-//
-//        let data = EstimateFeeRequestData(amount: amount.tokenValue,
-//                                          currency: amount.currency.code,
-//                                          destination: address)
-//
-//        EstimateFeeWorker().execute(requestData: data) { [weak self] result in
-//            switch result {
-//            case .success(let fee):
-//                completion(fee?.fee)
-//
-//            case .failure(let error):
-//                self?.presenter?.presentError(actionResponse: .init(error: error))
-//                completion(nil)
-//            }
-//        }
-//    }
 }
